requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-Login>=0.6
Flask-WTF>=1.1
Flask-SocketIO>=5.3
python-dotenv>=1.0
Flask-Bootstrap>=3.3
Flask-Mail>=0.9.1
Flask-Admin>=1.6
Flask-Cors>=4.0
Werkzeug>=2.2
Jinja2>=3.1
WTForms>=3.0
Faker>=19.0
plotly>=5.0
requests>=2.28 